+++ R/retrieving.R
@@ -3,6 +3,8 @@
 #' If `hashed_table` has `0` rows, `NULL` is returned.
 #'
 #' @param hashed_table A table created by `pl_hash()`.
+#' @param version_string A string of length 1 indicating the version to be downloaded.
+#'                       `NULL`, the default, means to download all versions.
 #' @param decode_fks A boolean that tells whether to decode foreign keys
 #'                   before returning.
 #'                   Default is `TRUE`.
@@ -31,7 +33,11 @@
 #' @param tar_group_colname The name of the `tar_group` column.
 #'                          default is `PFUPipelineTools::hashed_table_colnames$tar_group_colname`.
 #' @param matname_colname,matval_colname Names used for matrix names and matrix values.
-#'                                       Defaults are "matname" and "matval".
+#'                                       Defaults are
+#'                                       `PFUPipelineTools::mat_meta_cols$matname` and
+#'                                       `PFUPipelineTools::mat_meta_cols$matval`, respectively.
+#' @param valid_from_version_colname,valid_to_version_colname Names for columns containing version information.
+#'                                                            Defaults are
 #' @param conn The database connection.
 #' @param schema The database schema (a `dm` object).
 #'               Default calls `schema_from_conn()`, but
@@ -44,21 +50,24 @@
 #'                         Default calls `get_all_fk_tables()`.
 #'                         Needed only when `decode_fks = TRUE` (the default).
 #'                         If foreign keys are not being decoded,
-#'                         setting `NULL` may improve speed.
+#'                         setting to `NULL` may improve speed.
 #' @param .table_name_col,.nested_hash_col  See `PFUPipelineTools::hashed_table_colnames`.
 #'
 #' @return The downloaded data frame described by `hashed_table`.
 #'
 #' @export
 pl_collect_from_hash <- function(hashed_table,
+                                 version_string = NULL,
                                  decode_fks = TRUE,
                                  retain_table_name_col = FALSE,
                                  set_tar_group = TRUE,
                                  decode_matsindf = TRUE,
                                  matrix_class = c("Matrix", "matrix"),
                                  tar_group_colname = PFUPipelineTools::hashed_table_colnames$tar_group_colname,
-                                 matname_colname = "matname",
-                                 matval_colname = "matval",
+                                 matname_colname = PFUPipelineTools::mat_meta_cols$matname,
+                                 matval_colname = PFUPipelineTools::mat_meta_cols$matval,
+                                 valid_from_version_colname = PFUPipelineTools::version_cols$valid_from_version,
+                                 valid_to_version_colname = PFUPipelineTools::version_cols$valid_to_version,
                                  conn,
                                  schema = schema_from_conn(conn = conn),
                                  fk_parent_tables = get_all_fk_tables(conn = conn, schema = schema),
@@ -72,6 +81,7 @@ pl_collect_from_hash <- function(hashed_table,
                                                       fk_parent_tables = fk_parent_tables),
                                  .table_name_col = PFUPipelineTools::hashed_table_colnames$db_table_name,
                                  .nested_hash_col = PFUPipelineTools::hashed_table_colnames$nested_hash_colname) {
+
   matrix_class <- match.arg(matrix_class)
   if (nrow(hashed_table) == 0) {
     return(NULL)
@@ -84,6 +94,10 @@ pl_collect_from_hash <- function(hashed_table,
     unique()
   assertthat::assert_that(length(table_name) == 1,
                           msg = "More than 1 table received in pl_collect_from_hash()")
+
+  out <- dplyr::tbl(conn, table_name)
+
+  # Filter on any foreign keys
   filter_tbl <- hashed_table |>
     PFUPipelineTools::tar_ungroup() |>
     dplyr::select(!dplyr::all_of(c(.table_name_col, .nested_hash_col))) |>
@@ -93,12 +107,28 @@ pl_collect_from_hash <- function(hashed_table,
                conn = conn,
                schema = schema,
                fk_parent_tables = fk_parent_tables)
-  out <- dplyr::tbl(conn, table_name)
+
   if (ncol(filter_tbl) > 0) {
     out <- out |>
       # Perform a semi_join to keep only the rows in x that have a match in y
       dplyr::semi_join(filter_tbl, copy = TRUE, by = colnames(filter_tbl))
   }
+
+  # Filter on the version, if requested
+  if (!is.null(version_string)) {
+    # Figure out the index for the version of interest to us.
+    version_index <- encode_version_string(version_string = version_string,
+                                           version_colname = valid_from_version_colname,
+                                           table_name = table_name,
+                                           conn = conn,
+                                           schema = schema,
+                                           fk_parent_tables = fk_parent_tables)
+    # Filter the outgoing data frame according to the version_index
+    out <- out |>
+      dplyr::filter(.data[[valid_from_version_colname]] <= version_index) |>
+      dplyr::filter(.data[[valid_to_version_colname]] >= version_index)
+  }
+
   out <- out |>
     dplyr::collect()
   if (decode_fks) {
@@ -394,17 +424,3 @@ pl_filter_collect <- function(db_table_name,
 
   return(out)
 }
-
-
-# vars_in_dots <- function(enquos_dots) {
-#   enquos_dots |>
-#     # as.character(enquos_dots) turns the quosure into a string
-#     # in which table column names are prefixed by "~" and
-#     # terminated with a white space.
-#     as.character() |>
-#     # This pattern extracts for all characters between "~" and whitespace,
-#     # namely the column names to be decoded.
-#     stringr::str_extract("(?<=~)([^\\s]+)") |>
-#     # Make sure we have no duplicates.
-#     unique()
-# }