+++ NEWS.md
@@ -8,6 +8,10 @@ Cite all releases with doi [10.5281/zenodo.8226419](https://doi.org/10.5281/zeno
 which always resolves to the latest release.
 
 
+* `pl_collect_from_hash()` has new argument `version_string`
+  that provides capability to filter downloads by version.
+  The default value (`NULL`) means that all versions
+  should be downloaded.
 * `pl_collect_from_hash()` now more convenient
   with default arguments that pull values from `conn`.
   The behavior of `pl_collect_from_hash()` is now consistent with